User Profile Picture Instagram Firebase
Now that the necessary Firebase SDKs have been installed and the user authentication step is complete, let's dive into the code that allows more data persistence.  First we'll implement the feature that saves a username into Firebase Database.  Then we move onto using UIImagePickerController to allow the user selection of the profile image in our application.

Lastly, we upload the image into Firebase Storage, which conveniently gives us back a URL for retrieval later.  Finally we take this URL and save it onto our user along with their username.