User Profile Image View (+8min von neuem Video) Instagram Firebase
In this lesson, we'll log the user back in using the credentials we originally created in the first couple of episodes.  In order to do so, we'll create a separate login screen that only contains two input UITextFields. To login this user, we will be calling Sign In instead of the create user method.

Let's also go over a simple way to check for when the  key is pressed and used that as a signal to login our user.