Profile Header and Grid Cells Instagram Firebase
In this lesson, I'll dedicate some time to show you how you can render how our header using various tricks with UIStackView. Finally, when we're done, we'll take a look at how to create the grid of cells for our UserProfileController.  Some bit of calculation regarding the pixel spacing is necessary to arrive at the correct layout of cells.