Logging Out with UIAlertController Instagram Firebase
To further complete our profile page, we'll implement the feature that allows the user to log out and log back in.  Let's use the UIAlertController Action Sheet component to present the necessary options to our user.  Once we execute a proper sign out, I'll demonstrate a technique that prompts the user to log back in right away when the app starts.