Login Design Instagram Firebase
Now that our LoginController is loaded up in our application, it's the perfect time to give it some nice design.  First we start by centering the Instagram logo at the very top with a blue background behind it. Then we borrow the UITextField and UIButton components from our SignUpController.

Lastly, I want to introduce the stylized attributed text for our button at the very bottom.  While we're at it, let's also use a similar button in SignUpController to pop back to our login screen.