Log In and UI Reset Instagram Firebase
In this lesson, we fix up the remaining components of our LoginController to allow the user the ability to log back into the application.  In addition to this, we'll inspect the behavior of our application after we've dismiss the login and registration screens.  To fix the issues that arise from logging in our users, we'll implement a quick fix that will reset the UI for our MainTabBarController.