Tricks to Custom Tab Bar Instagram Firebase
Making our application's nice and pretty really sets it apart from the competition.  In this lesson, let's add in the four additional icons to the bottom UITabBar of our main UI.  Along the way, I'll show you some neat tricks that make this process a lot easier through an Optional Method Parameter.  Finally, once our buttons are set up, we center the icons with a couple of inset parameters.