Present Photo Selector Instagram Firebase
We are now ready to allow the user to start adding in photos to the application by using the plus button in the middle of the Tab Bar.  We'll look at how to custom the behavior of UITabBarController to handle pressing of a button and disabling the default behavior.  Furthermore, our logic will present an image picker controller that the user can use for photo selection.