User Photo Selection Instagram Firebase
In this lesson, we'll be implementing didSelectItem in our UICollectionView to handling the selection of our photos in the list.  In order to show the selected photo, we will setup a custom header that holds reference to a UIImageView that we'll set using our selected image.

In addition to this, we will analyze the behavior of efficiency and UI hanging when trying to optimize for the best user experience.