Share Photo Caption Instagram Firebase
With a selected photo now available in our app, let's work on the next page where a user fills out the caption of their post.  We'll start by adding a top section for the selected image and caption text using a UIImageView and UITextView.

Next we need to somehow send in the selected image from the previous screen into this UIImageView. In order to do this, we'll look back into how our large size header image is being rendered out and add in some logic.