User Search Controller Instagram Firebase
Our world is currently quite empty because we don't have any way of searching for users in the application.  In this lesson, let's go ahead and create a few additional users in our system.  Next, we'll set up the search screen that allows easy filtering of users based on the text input in the navigational search bar.