Search Filtering Instagram Firebase
In order to perform a fast and efficient search for the users in our Firebase Database, we'll first need to fetch all of our users with an observation.  Once we build our list of users, we'll present it inside of our UserSearchController.  Next, we'll examine the behavior of delegating our search bar behavior to detect when the search text is changed.  Finally, we can perform a filtering of our users so that our list responds when searching.