Following and Unfollowing Instagram Firebase
This lesson will cover the aspects of following and unfollowing users in the system.  More specifically, we'll make a server call to Firebase Database that will create a branch for following users.  Using this branch as our setup, we'll be able to figure out later how to piece together our home feed with all of the necessary posts.

Don't forget, once we follow a user, we'll also need to unfollow them.  This will be covered in addition to styling our button based on the follow relationship of our users.