Following Users Posts Instagram Firebase
Now that we the capability of following users inside of our system, it would be nice if we improve our Home Feed to show their posts.  In today's video, I'll go through a few examples of how this works by fetching posts for various users in the system and rendering it in our list.

Finally, we go over how to sort our posts in a descending date order so we can see the newest items at the top.  Once we complete this feature, I'll pull in an extension method on Date that returns a timeAgoDisplay string for our posts.