Manual Automatic Post Refresh Instagram Firebase
In today's lesson, we're going to cover the topic of implementing an easy way to refresh our feed using a UIRefreshControl.  This component allows us to present a UIActivityIndicatorView when we drag our UICollectionView down to refresh.

Next, we'll look at how to automatically refresh our home feed every time we upload a new post to Firebase Database.  For this to work, we'll look at implementing NotificationCenter observers.