Comments Custom Delegation Instagram Firebase
Let's now move onto the beginning of our implementation for showing Comments inside of our application. First we'll look at how to implement Custom Delegation that allows us to tap on an icon in a cell and propagate it up to the controller.  Furthermore, we will discuss how to find out which particular list item you are clicking on.