Comments Input Accessory View Instagram Firebase
Before we can render out our comments inside of the UICollectionView, we'll need to implement a section that contains the input components necessary.  In this lesson, I'll go over a very simple way to implement a keyboard handler by using an inputAccessoryView property.  Using this approach will allow the list above to interact with the input components naturally and easily.