Insert Comments Instagram Firebase
Now that our interface is ready for text input, let's begin the process of inserting comments for the user that is currently logged in.  We'll be appending onto a comments node in Firebase to make the retrieval fast and simple.  The three properties we'll need to keep track of are the comment text, creationDate, and user id.