Fetch Comments Instagram Firebase
With the list of comments now inserted into Firebase Database, we're ready to render them out into our UICollectionView.  We'll first start by laying out our CommentCell views in our controller. Next, we perform the fetching of comment based on the post that we are tapping into.  Performing all of this logic will get us ready for the next lesson on further details of each cell.